requirements.txt
kafka-python==2.0.2
requests>=2.32.3,<3.0.0
pymysql
cryptography
psycopg2-binary
